refactor: [#1189] new whitelist::repository module
