+++ src/core/whitelist/repository/mod.rs
@@ -1,0 +1,2 @@
+pub mod in_memory;
+pub mod persisted;
